chore(wallets): extract RenameWallet into its own component
