+++ extension/src/popup/views/Wallets/styles.scss
@@ -145,25 +145,3 @@
   background-color: var(--sds-clr-gray-01);
   z-index: var(--z-index--banner);
 }
-
-.RenameWallet {
-  z-index: var(--z-index--banner);
-  position: relative;
-
-  &__short-address {
-    font-size: 14px;
-    color: var(--sds-clr-gray-10);
-    margin: #{pxToRem(12px)} 0 #{pxToRem(18px)} 0;
-  }
-
-  &__actions {
-    display: flex;
-
-    button {
-      flex: 1;
-    }
-    button:not(:first-child) {
-      margin-left: #{pxToRem(10px)};
-    }
-  }
-}
